Finish off create, read, update and delete for directories

+++ handlers.go
@@ -28,7 +28,7 @@ func handleReadDir(w http.ResponseWriter, req *http.Request) {
 	}
 
 	// check that directory exists
-	if dirExists(fp) {
+	if dirExists(fp) == false {
 		errDirNotFound.Write(w)
 		return
 	}
@@ -55,30 +55,36 @@ func handleCreateDir(w http.ResponseWriter, req *http.Request) {
 		return
 	}
 
+	// combine parent dir with dir name
 	dirname := req.FormValue("dir[name]")
 	fp := filepath.Join(parent, dirname)
 
+	// check if dir already exists
 	if dirExists(fp) {
 		errDirConflict.Write(w)
 		return
 	}
 
+	// make directory
 	if err = os.MkdirAll(fp, 0755); err != nil {
 		NewApiError(err).Write(w)
 		return
 	}
 
+	// check that directory was created
 	info, err := os.Stat(fp)
 	if err != nil {
 		NewApiError(err).Write(w)
 		return
 	}
 
+	// convert fileinfo into something we can print
 	dir := &rangolib.File{
 		Path: strings.TrimPrefix(fp, CONTENT_DIR),
 	}
 	dir.Load(info)
 
+	// print info
 	printJson(w, struct {
 		Dir *rangolib.File `json:"dir"`
 	}{
@@ -88,24 +94,83 @@ func handleCreateDir(w http.ResponseWriter, req *http.Request) {
 
 // handleUpdateDir renames a directory
 func handleUpdateDir(w http.ResponseWriter, req *http.Request) {
+	fp, err := sanitizePath(mux.Vars(req)["path"])
+	if err != nil {
+		errInvalidDir.Write(w)
+		return
+	}
+
+	// check that the specified directory is not the root content folder
+	if fp == CONTENT_DIR {
+		errInvalidDir.Write(w)
+		return
+	}
+
+	// check that directory exists
+	if dirExists(fp) == false {
+		errDirNotFound.Write(w)
+		return
+	}
+
+	// combine parent dir with dir name
+	parent := filepath.Dir(fp)
+	dirname := req.FormValue("dir[name]")
+	dest := filepath.Join(parent, dirname)
+
+	// rename directory
+	if err = os.Rename(fp, dest); err != nil {
+		NewApiError(err).Write(w)
+		return
+	}
+
+	// check that directory was created
+	info, err := os.Stat(dest)
+	if err != nil {
+		NewApiError(err).Write(w)
+		return
+	}
+
+	// convert fileinfo into something we can print
+	dir := &rangolib.File{
+		Path: strings.TrimPrefix(dest, CONTENT_DIR),
+	}
+	dir.Load(info)
+
+	// print info
+	printJson(w, struct {
+		Dir *rangolib.File `json:"dir"`
+	}{
+		Dir: dir,
+	})
 }
 
 // handleDeleteDir deletes a directory
 func handleDeleteDir(w http.ResponseWriter, req *http.Request) {
 	fp, err := sanitizePath(mux.Vars(req)["path"])
 	if err != nil {
-		fmt.Fprint(w, err)
+		errInvalidDir.Write(w)
 		return
 	}
 
+	// check that the specified directory is not the root content folder
+	if fp == CONTENT_DIR {
+		errInvalidDir.Write(w)
+		return
+	}
+
+	// check that directory exists
+	if dirExists(fp) == false {
+		errDirNotFound.Write(w)
+		return
+	}
+
+	// remove the directory
 	if err = os.RemoveAll(fp); err != nil {
-		fmt.Fprint(w, err)
+		NewApiError(err).Write(w)
 		return
 	}
 
-	printJson(w, map[string]interface{}{
-		"path": fp,
-	})
+	w.WriteHeader(http.StatusNoContent)
 }
 
 //   __        __   ___  __
